Final Fix
Separated image model code from scraper.py
Removed sessionid
